pyproject.toml
[project]
name = "payd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.0"
dependencies = [
    "django>=6.0",
    "django-ninja>=1.5.1",
    "google>=3.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
